requirements.txt
Django==5.1.7
python-dotenv==1.0.1
python-telegram-bot==22.0
ruff==0.11.0
psycopg2